// packages/frontend/src/generated/types/TruckWaybillResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `truck_waybill` entity.
 */
export interface TruckWaybillResponse {
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string
   */
  date: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  senderId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}